requirements.txt
numpy==1.26.4
opencv-python-headless==4.11.0.86
Pillow==10.2.0
moderngl[headless]
pyrr
